+++ src/component/TOS/TOSPage.tsx
@@ -1,4 +1,4 @@
-import React, { cloneElement, ReactElement, useState } from "react";
+import React, { cloneElement, ReactElement } from "react";
 import { ModalProps, Stack, StackProps } from "@mantine/core";
 import { mergeSx } from "../../util";
 import { useDisclosure } from "@mantine/hooks";