+++ src/component/Theme/Paper/ThemePaper.tsx
@@ -1,12 +1,12 @@
 import React from "react";
-import { useMantineColorScheme, SharedPaperProps } from "@mantine/core";
+import { useMantineColorScheme, PaperProps } from "@mantine/core";
 import { DarkPaper } from "./DarkPaper";
 import { LightPaper } from "./LightPaper";
 
 /**
  * A {@link https://mantine.dev/core/paper Paper} component that automatically changes with the theme.
  */
-export function ThemePaper({ ...other }: SharedPaperProps) {
+export function ThemePaper({ ...other }: PaperProps) {
     const { colorScheme } = useMantineColorScheme();
 
     return (