+++ src/util/interface/caption/SharedCaptionProps.ts
@@ -1,0 +1,3 @@
+import { TextProps } from "@mantine/core";
+
+export interface SharedCaptionProps extends TextProps {}